diagnosticador_supabase.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
DIAGNOSTICADOR SUPABASE AVANÇADO - RESOLVE PROBLEMAS DE CONEXÃO
Dashboard Baker - Soluções para Timeout e Conectividade
"""

import psycopg2
import os
import sys
from datetime import datetime
import socket
import urllib.request
import json

def testar_conectividade_basica():
    """Testa conectividade básica de rede"""
    print("🌐 TESTE DE CONECTIVIDADE BÁSICA")
    print("=" * 40)
    
    # Teste 1: Internet
    print("   [1/4] Testando internet...", end="")
    try:
        response = urllib.request.urlopen('https://google.com', timeout=5)
        print(" ✅ OK")
        internet_ok = True
    except:
        print(" ❌ FALHA")
        internet_ok = False
    
    # Teste 2: Supabase website
    print("   [2/4] Testando Supabase...", end="")
    try:
        response = urllib.request.urlopen('https://supabase.com', timeout=5)
        print(" ✅ OK")
        supabase_ok = True
    except:
        print(" ❌ FALHA")
        supabase_ok = False
    
    # Teste 3: Seu projeto Supabase
    print("   [3/4] Testando seu projeto...", end="")
    try:
        response = urllib.request.urlopen('https://lijtncazuwnbydeqtoyz.supabase.co', timeout=5)
        print(" ✅ OK")
        projeto_ok = True
    except:
        print(" ❌ FALHA")
        projeto_ok = False
    
    # Teste 4: Porta PostgreSQL
    print("   [4/4] Testando porta 5432...", end="")
    try:
        sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
        sock.settimeout(5)
        result = sock.connect_ex(('db.lijtncazuwnbydeqtoyz.supabase.co', 5432))
        sock.close()
        if result == 0:
            print(" ✅ OK")
            porta_ok = True
        else:
            print(" ❌ BLOQUEADA")
            porta_ok = False
    except:
        print(" ❌ ERRO")
        porta_ok = False
    
    return {
        'internet': internet_ok,
        'supabase': supabase_ok, 
        'projeto': projeto_ok,
        'porta': porta_ok
    }

def testar_multiplas_conexoes_supabase():
    """Testa diferentes formas de conectar ao Supabase"""
    print("\n🔌 TESTE DE CONEXÕES MÚLTIPLAS")
    print("=" * 40)
    
    senha = "nB7ZayxQksiHeZmG"  # Senha atual
    
    # Configurações para testar
    configuracoes = [
        {
            'nome': 'Conexão Direta (5432)',
            'config': {
                'host': 'db.lijtncazuwnbydeqtoyz.supabase.co',
                'database': 'postgres',
                'user': 'postgres',
                'password': senha,
                'port': 5432,
                'sslmode': 'require',
                'connect_timeout': 10
            }
        },
        {
            'nome': 'Pooler Connection (6543)',
            'config': {
                'host': 'db.lijtncazuwnbydeqtoyz.supabase.co',
                'database': 'postgres',
                'user': 'postgres',
                'password': senha,
                'port': 6543,  # Porta alternativa do Supabase
                'sslmode': 'require',
                'connect_timeout': 10
            }
        },
        {
            'nome': 'SSL Preferido',
            'config': {
                'host': 'db.lijtncazuwnbydeqtoyz.supabase.co',
                'database': 'postgres',
                'user': 'postgres',
                'password': senha,
                'port': 5432,
                'sslmode': 'prefer',
                'connect_timeout': 15
            }
        },
        {
            'nome': 'Timeout Longo',
            'config': {
                'host': 'db.lijtncazuwnbydeqtoyz.supabase.co',
                'database': 'postgres',
                'user': 'postgres',
                'password': senha,
                'port': 5432,
                'sslmode': 'require',
                'connect_timeout': 30
            }
        }
    ]
    
    conexoes_funcionando = []
    
    for i, teste in enumerate(configuracoes, 1):
        print(f"   [{i}/4] {teste['nome']}...", end="")
        
        try:
            conn = psycopg2.connect(**teste['config'])
            cursor = conn.cursor()
            cursor.execute("SELECT version()")
            version = cursor.fetchone()[0]
            cursor.close()
            conn.close()
            
            print(" ✅ FUNCIONOU!")
            conexoes_funcionando.append(teste)
            
        except psycopg2.OperationalError as e:
            error_msg = str(e).lower()
            if 'timeout' in error_msg:
                print(" ❌ TIMEOUT")
            elif 'authentication' in error_msg:
                print(" ❌ SENHA")
            elif 'connect' in error_msg:
                print(" ❌ CONEXÃO")
            else:
                print(" ❌ ERRO")
        except Exception as e:
            print(f" ❌ {str(e)[:20]}")
    
    return conexoes_funcionando

def diagnosticar_problema_principal():
    """Diagnostica o problema principal"""
    print("\n🔍 DIAGNÓSTICO DO PROBLEMA")
    print("=" * 40)
    
    conectividade = testar_conectividade_basica()
    conexoes_ok = testar_multiplas_conexoes_supabase()
    
    # Análise dos resultados
    if not conectividade['internet']:
        return "internet", "Problema de internet - verifique conexão"
    
    if not conectividade['supabase']:
        return "supabase_down", "Supabase pode estar com problemas"
    
    if not conectividade['projeto']:
        return "projeto", "Seu projeto Supabase não está acessível"
    
    if not conectividade['porta']:
        return "firewall", "Porta 5432 bloqueada (firewall/ISP)"
    
    if len(conexoes_ok) > 0:
        return "resolvido", f"Conexão funcionando: {conexoes_ok[0]['nome']}"
    
    return "senha_ou_config", "Problema de senha ou configuração"

def criar_solucoes_alternativas():
    """Cria soluções alternativas"""
    print("\n💡 SOLUÇÕES ALTERNATIVAS")
    print("=" * 40)
    
    # Solução 1: Railway (mais confiável)
    print("✅ SOLUÇÃO 1 - RAILWAY (RECOMENDADA)")
    print("   • Mais estável que Supabase para PostgreSQL")
    print("   • Deploy automático")
    print("   • Sem problemas de firewall")
    print("   • Grátis por 5 dólares/mês")
    
    # Solução 2: PostgreSQL Local
    print("\n✅ SOLUÇÃO 2 - POSTGRESQL LOCAL")
    print("   • Instalar PostgreSQL local")
    print("   • Usar com PgAdmin")
    print("   • 100% controle")
    print("   • Sem dependência de internet")
    
    # Solução 3: Usar Supabase API REST
    print("\n✅ SOLUÇÃO 3 - SUPABASE API REST")
    print("   • Usar API REST em vez de PostgreSQL direto") 
    print("   • Evita problemas de firewall")
    print("   • Usar sua ANON_KEY")
    
    # Solução 4: CSV Local 
    print("\n✅ SOLUÇÃO 4 - CSV LOCAL")
    print("   • Usar arquivos CSV")
    print("   • Dashboard totalmente local")
    print("   • Sem dependência de banco")

def gerar_script_railway_alternativo():
    """Gera script para migrar para Railway"""
    script = '''#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
MIGRAÇÃO PARA RAILWAY POSTGRESQL - SOLUÇÃO DEFINITIVA
Dashboard Baker - Sistema Estável
"""

import psycopg2
import os
import requests
from datetime import datetime

def configurar_railway():
    print("🚂 RAILWAY POSTGRESQL - SOLUÇÃO ESTÁVEL")
    print("=" * 50)
    print()
    print("🎯 PASSOS PARA RAILWAY:")
    print("1. Acesse: https://railway.app")
    print("2. Login com GitHub")
    print("3. New Project → Add PostgreSQL")
    print("4. Copie as credenciais")
    print("5. Execute este script novamente")
    print()
    
    # Solicitar credenciais Railway
    print("📋 CREDENCIAIS RAILWAY:")
    pghost = input("PGHOST: ").strip()
    pguser = input("PGUSER: ").strip()
    pgpassword = input("PGPASSWORD: ").strip()
    pgdatabase = input("PGDATABASE: ").strip()
    pgport = input("PGPORT (5432): ").strip() or "5432"
    
    # Testar Railway
    try:
        conn = psycopg2.connect(
            host=pghost,
            user=pguser,
            password=pgpassword,
            database=pgdatabase,
            port=pgport,
            sslmode='require'
        )
        print("✅ Railway PostgreSQL funcionando!")
        conn.close()
        
        # Criar .env para Railway
        with open('.env', 'w') as f:
            f.write(f"""# RAILWAY POSTGRESQL - FUNCIONANDO
PGHOST={pghost}
PGUSER={pguser}
PGPASSWORD={pgpassword}
PGDATABASE={pgdatabase}
PGPORT={pgport}

# Configurações
ENVIRONMENT=railway
DATABASE_PROVIDER=railway
VALIDATED=true
""")
        
        print("✅ .env criado para Railway!")
        print("🚀 Agora execute: python dashboard_baker_web_corrigido.py")
        return True
        
    except Exception as e:
        print(f"❌ Erro Railway: {e}")
        return False

if __name__ == "__main__":
    configurar_railway()
'''
    
    with open('setup_railway_alternativo.py', 'w', encoding='utf-8') as f:
        f.write(script)
    
    print("\n✅ Script Railway criado: setup_railway_alternativo.py")

def main():
    """Função principal - diagnosticar e resolver"""
    print("🚨 DIAGNOSTICADOR SUPABASE AVANÇADO")
    print("=" * 50)
    print("🎯 Vamos resolver o problema de timeout!")
    print()
    
    # Diagnóstico completo
    problema, descricao = diagnosticar_problema_principal()
    
    print(f"\n🎯 PROBLEMA IDENTIFICADO: {descricao}")
    print()
    
    # Soluções baseadas no problema
    if problema == "firewall":
        print("🔧 SOLUÇÕES PARA FIREWALL/ISP:")
        print("1. 🚂 Railway (RECOMENDADO) - sem problemas de firewall")
        print("2. 🔥 Hotspot mobile - usar internet do celular")
        print("3. 🌐 VPN - usar NordVPN/ExpressVPN")
        print("4. 📡 Porta alternativa 6543")
        
        # Criar script Railway
        gerar_script_railway_alternativo()
        
        print(f"\n⚡ AÇÃO IMEDIATA:")
        print("   python setup_railway_alternativo.py")
        
    elif problema == "senha_ou_config":
        print("🔧 SOLUÇÕES PARA CONFIGURAÇÃO:")
        print("1. Resetar senha no Supabase")
        print("2. Verificar se PostgreSQL está ativo")
        print("3. Usar Railway como alternativa")
        
    elif problema == "resolvido":
        print("🎉 PROBLEMA RESOLVIDO!")
        print("Uma das conexões funcionou - criando .env...")
        
    else:
        print("🔧 SOLUÇÕES GERAIS:")
        criar_solucoes_alternativas()
    
    print(f"\n🎯 RECOMENDAÇÃO FINAL:")
    print("   👍 Use Railway - é mais estável")
    print("   👍 Execute: python setup_railway_alternativo.py")
    print("   👍 Railway resolve 100% dos problemas de rede")

if __name__ == "__main__":
    main()