requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.17.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
xlsxwriter>=3.1.0
numpy>=1.24.0
